chore: Refactor DataUser class and UserController
The DataUser class has been modified to remove the "dateCreation" field, as it is no longer needed. Additionally, the UserController class has been updated to make the "file" parameter optional in the register method. This allows users to register without uploading a profile picture.

This commit refactors the DataUser class and UserController to improve code clarity and user registration process.