chore: Add DataUserLogin class and login endpoint
This commit adds the `DataUserLogin` class to handle the login data of a user. It also includes the implementation of the login endpoint in the `UserController` class. The endpoint receives the user login data in the request body and authenticates the user using the provided credentials. If the authentication is successful, it returns a JWT token. This addition enhances the user authentication functionality of the application.